ESP32-C3 RC

+++ ESP-IDF_Robot_RC/main/espnow_example_main.c
@@ -467,8 +467,8 @@ void sendData (void) {
     buffer = malloc(sizeof(sensors_data_t));
     buffer->type = 1;
     buffer->crc = 0;
-    buffer->x_axis = (uint8_t)240;
-    buffer->y_axis = (uint8_t)2040;
+    buffer->x_axis = 240;
+    buffer->y_axis = 2040;
     buffer->nav_bttn = 0;
     buffer->motor1_rpm_pcm = 10;
     buffer->motor2_rpm_pcm = 0;
